Resolved compilation and runtime errors
