+++ CMP1903M A01 2223/Card.cs
@@ -12,7 +12,45 @@ namespace CMP1903M_A01_2223
         //Value: numbers 1 - 13
         //Suit: numbers 1 - 4
         //The 'set' methods for these properties could have some validation
-        public int Value { get; set; }
-        public int Suit { get; set; }
+        private int _value;
+        private int _suit;
+        public int Value
+        {
+            get { return _value; }
+            set
+            {
+                bool validVal = false;
+                if (Value > 0)
+                {
+                    if (Value < 14)
+                    {
+                        validVal = true;
+                    }
+                }
+                _value = Value;
+            }
+        }
+        public int Suit
+        {
+            get { return _suit; }
+            set
+            {
+                bool validSuit = false;
+                if (Suit > 0)
+                {
+                    if (Suit < 14)
+                    {
+                        validSuit = true;
+                    }
+                }
+                _suit = Suit;
+            }
+        }
+        //Card constructor
+        public Card(int value, int suit)
+        {
+            Value = value;
+            Suit = suit;
+        }
     }
 }
